Added protection when destroying `context_refreshing_timer'.
git-svn-id: https://svn.eiffel.com/eiffelstudio-public/trunk@55769 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/Eiffel/interface/new_graphical/windows/eb_development_window.e
@@ -3539,7 +3539,10 @@ feature {NONE} -- Implementation
 				end
 			else
 				Precursor {EB_TOOL_MANAGER}
-				context_refreshing_timer.destroy
+				if context_refreshing_timer /= Void then
+					context_refreshing_timer.destroy
+					context_refreshing_timer := Void
+				end
 			end
 		end
